+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.0.27] and [v0.0.26] November 28th, 2023.
+
+### Added
+
+- Now uses `#` anchors to directly get to a state. Also, send that to a parent
+  window, if any, and listen for messages from the parent window for state to
+  jump into.
+
 ## [v0.0.25] and [v0.0.26] November 20th, 2023.
 
 ### Added